Move search command to different extension cog
